requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.12.0
mccabe==0.7.0
pep8-naming==0.13.3
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
